occasion/pyproject.toml
[tool.poetry]
name = "occasion"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-enumfields = "^2.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
